Subtle change, big difference: Removed all occurences of NeXT_RUNTIME and instead used GNU_RUNTIME where appropriate. Why did I do this? Because of precompiled headers! Everyone using precompiled headers had to define NeXT_RUNTIME in his projects to access the NeXT_RUNTIME definitions correctly. Because the headers were never precompiled with this flag, they were turned down during compiling. Re-precompiling them with the correct settings however required every client project - even those that didn't care - to define NeXT_RUNTIME; if they didn't they couldn't use the precompiled headers as well ... grmpf.
I figured, that the GNUstep platform doesn't have precompiled headers and doesn't suffer from any of the above mentioned problems. So I defined GNU_RUNTIME in all GNUmakefiles - and it turns out to be even better, because it's easy to figure in the GNUstep make environment, if you REALLY are using the GNU runtime.

Classical win-win here. :-)