Fixed yet another bug in the EDMLParser. The following is valid XML:
<!DOCTYPE rdf:RDF [
  <!ENTITY % HTMLlat1 PUBLIC
  "-//W3C//ENTITIES Latin 1 for XHTML//EN"
  "http://www.w3.org/TR/xhtml1/DTD/xhtml-lat1.ent">
  %HTMLlat1;
]>

The new code that deals with <![CDATA[ ... ]]> correctly would let this slip through. I added a new EDMLPCommentBracketMode which is aware of <![CDATA[ also and will branch into EDMLPCDATAMode when this is appropriate.

I tested this with a variety of RSS/RDF feeds. This doesn't guarantee that it's 100% bugfree now, though.